Extract to separate import
